Add cancel button that cancels that prompts the user and the cancels the task
