+++ backbone/src/main/res/values/colors.xml
@@ -30,4 +30,6 @@
     <color name="rsb_forgot_passcode_text_color">@color/rsb_submit_bar_negative</color>
     <color name="rsb_fingerprint">@color/rsb_colorPrimary</color>
 
+    <color name="rsb_cancel_menu_icon">@color/rsb_white</color>
+
 </resources>
